Refactor invitation email component structure
- Moved InvitationEmail component to a new location under @lib/email/templates.
- Updated import paths in the invitation service to reflect the new component structure.
